refactor: update Genre import statements to use enums for consistency
