+++ src/main/java/br/com/rodrigo/onlinelibraryapi/repositories/BookRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import br.com.rodrigo.onlinelibraryapi.entities.Book;
-import br.com.rodrigo.onlinelibraryapi.entities.Genre;
+import br.com.rodrigo.onlinelibraryapi.enums.Genre;
 import jakarta.transaction.Transactional;
 
 @Repository
