+++ src/test/java/br/com/rodrigo/onlinelibraryapi/repositories/BookRepositoryTest.java
@@ -2,7 +2,7 @@ package br.com.rodrigo.onlinelibraryapi.repositories;
 
 import br.com.rodrigo.onlinelibraryapi.entities.Author;
 import br.com.rodrigo.onlinelibraryapi.entities.Book;
-import br.com.rodrigo.onlinelibraryapi.entities.Genre;
+import br.com.rodrigo.onlinelibraryapi.enums.Genre;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
